Fix missing-value leakage in ISMIP7 atmosphere forcing; share extrapolation helper
process_smb.py, process_temperature.py, process_smb_gradient.py, and
process_temperature_gradient.py remapped the raw source file directly with
ncremap, unlike process_runoff.py, ocean_thermal, and the fracture pathway,
which all extrapolate fill/missing values on the source grid first. Without
that step, no-data cells near the ice-sheet margin (e.g. ocean cells in the
RACMO OCX source) leak the ~9.97e36 netCDF fill sentinel into valid
neighboring cells during conservative/bilinear remapping.

Add the same pre-remap extrapolation to the four affected steps.

Also promote the previously-duplicated extrapolate_source() (from
fracture/remap_utils.py, formerly re-implemented again in
ocean_thermal/process_thermal_forcing.py and process_runoff.py) to a single
shared compass.landice.tests.ismip7_forcing.remap_utils module, and update
all five call sites to use it instead of per-file private copies.
