+++ compass/landice/tests/ismip7_forcing/atmosphere/process_temperature_gradient.py
@@ -10,6 +10,7 @@ from compass.landice.tests.ismip7_forcing.create_mapfile import (
     build_mapping_file,
 )
 from compass.landice.tests.ismip7_forcing.ice_sheet_params import get_params
+from compass.landice.tests.ismip7_forcing.remap_utils import extrapolate_source
 from compass.step import Step
 
 
@@ -128,14 +129,23 @@ class ProcessTemperatureGradient(Step):
                 logger.info(f"  Remapped file exists, skipping: {basename}")
                 continue
 
+            # Extrapolate fill values on source grid before remapping
+            # so they don't pollute neighboring cells during interpolation
+            extrap_file = f"extrap_{basename}"
+            if not os.path.exists(extrap_file):
+                extrapolate_source(input_file, extrap_file, "dtsdz", logger)
+
             logger.info(f"  Remapping: {basename}")
             args = ["ncremap",
-                    "-i", input_file,
+                    "-i", extrap_file,
                     "-o", remapped_file,
                     "-m", mapping_file,
                     "-v", "dtsdz"]
 
             check_call(args, logger=logger)
+
+            # Clean up extrapolated source file
+            os.remove(extrap_file)
 
         # Combine remapped files and rename to MALI conventions
         logger.info("Combining remapped files and renaming variables...")
